+++ test/core/flat_buffer.cpp
@@ -12,7 +12,7 @@
 
 #include <beast/core/ostream.hpp>
 #include <beast/core/read_size.hpp>
-#include <beast/core/string_view.hpp>
+#include <beast/core/string.hpp>
 #include <beast/test/test_allocator.hpp>
 #include <beast/unit_test/suite.hpp>
 #include <boost/lexical_cast.hpp>